Update the FAQ.
Is there anyone who could give me an account on a faster-than-20-bytes-per
second line to the US (not EBONE) ? I'm having too many problems connecting
to freefall to do commits... It would be better if it could be a ssh-protected
account, that would give a strong authentication.

The RENATER line (and thus EBONE) is almost dead for remote login after
10AM european time... :-(

Thanks in advance.


git-svn-id: https://svn.freebsd.org/doc/head@145 c2e8774f-c49f-e111-b436-862b2bbc8956