Add an entry about slow starting SCSI CD-ROM:
When I booted my system without the above option, the CDROM could not
respond in time to the bootup probe of devices and was "missed".  When
I tried to access the device I got the "Device not configured" error
message.  I rebuilt the kernel with the SCSI_DELAY option and the
problem went away.

Submitted by:	Jon Wallace <adrl@whoweb.com>


git-svn-id: https://svn.freebsd.org/doc/head@610 c2e8774f-c49f-e111-b436-862b2bbc8956