In the process of re-constructing the FreeBSD web site automatic-update, clean up the build process a wee bit.
The basic change: instead of cheacking out bits and pieces of the
doc tree into the web build tree, check out the doc tree somewhere
else and put in a few symlinks from the web tree to the doc tree.

On catfish (soon to be hub) for example:

  /usr/local/www/build/doc/...
  /usr/local/www/build/www/...

where

  /usr/local/www/build/www/data/handbook -> ../../doc/handbook

You have to manually put in the symlinks at the moment, but it
works better than the evil CVS hackery.

Also, install with -C to help avoid gratuitous cache-busting due to
gratuitous timestamp twiddling.


git-svn-id: https://svn.freebsd.org/doc/head@2376 c2e8774f-c49f-e111-b436-862b2bbc8956