Changed the construct
    <li>....<p></p></li>

which was used all over the file to

    <li><p>...</p></li>

For all entries other than the ones in July 1998 there's no user visible
effect. For the entries in July (which were missing the <p></p>) this has
the effect (on some browsers at least) of inserting white space between
the list items, which is consistent with the appearance of the rest of the
file.


git-svn-id: https://svn.freebsd.org/doc/www@3080 c2e8774f-c49f-e111-b436-862b2bbc8956