Removed double whitespace from the end of stops -- . , ! ? : ;
Some parts of the handbook had single spaces after stops, some had double
or triple. While the typographical convention for monospaced fonts may
be to use double spaces after them, that doesn't apply here. TeX will
ignore them, as will HTML. If we need them for a plain text version of the
Handbook then the stylesheet / conversion mechanism can insert them
as necessary.

Searching for

   _\([;:!\.\?,]\)  +_

in Emacs and replacing with

   _\1 _

(ignore the '_', they're just to delineate the regexps) does the job
quite nicely. However, you can't do this everywhere, since some of the
double spaces might be in program listings or other literal sections
(e.g., the BSD Copyright), so you need to sit and bounce on the 'y' or
'n' key as appropriate for each occurance of a stop.


git-svn-id: https://svn.freebsd.org/doc/head@3646 c2e8774f-c49f-e111-b436-862b2bbc8956