With the agreement of the Japanese team, a change in the way I'm doing things.
I'm now working through from the beginning of the handbook to end,
correcting as I go. I'll commit in chunks of 5,000 lines (or
thereabouts).

Most of the changes fall into the following categories.

  * <emphasis remap=bf> --> <emphasis>

  * Spurious <para>s around <*list>s deleted (but not reformatted)

    "C-c -" in Emacs SGML mode (when the point is on an element starting
    or end tags) will delete that element's starting or end tags.

  * Marked smileys with <!-- smiley --> for possible future deletion

  * Deleting <emphasis>, around

    <term><emphasis>...</emphasis></term> -> <term>...</term>

  * Fine tuning markup choices in some cases

    - <filename>C:</filename>  ->  <devicename>C:</devicename>

  * Extra <note>s here and there.

  * Some <*list>s to <procedure> (and <listitem>s to <step>)

  * ASCII emphasis converted to <emphasis>

    i.e.,   do it like *this*  ->  do it like <emphasis>this</emphasis>

  * <symbol> -> <replaceable>

There are very few whitespace changes, although a few have probably
cropped up. The vast majority of the whitespace changes will happen in
one megacommit, hopefully some time next week.

This does the first 5,000 lines or so.


git-svn-id: https://svn.freebsd.org/doc/head@3649 c2e8774f-c49f-e111-b436-862b2bbc8956