Added output.html and output.print as parameter entities to freebsd.dtd. Default to "IGNORE", you should set them to "INCLUDE" by whatever mechanism your SGML parser supports as necessary (i.e., Jade's -i flag).
Added redefinitions for some ISO entities that aren't understood by
most web browsers. Since this is for HTML output only, wrap it in a
%output.html; marked section.

Use these marked sections to put both print and HTML stylesheets in one
.dsl file (freebsd.dsl). The print stylesheet now understands about the
elements that have been added to DocBook (and won't try and render them
in red).

Updated Makefile to use output.html and output.print on the command
line to Jade, instead of the earlier "html" and "print.

Note: producing .tex (and thence .ps and .pdf) versions of the Handbook
is broken. The Handbook tickles several bugs in the JadeTeX macros.
Sebastian Rahz, the JadeTeX author, knows about this, and is working on
fixing them.


git-svn-id: https://svn.freebsd.org/doc/head@3950 c2e8774f-c49f-e111-b436-862b2bbc8956