Created a new set of entities for manual pages.
The construct:

    <citerefentry>
      <refentrytitle>foobar</refentrytitle>
      <manvolnum>1</manvolnum>
    </citerefentry>

is a pain to type, and messes up the pretty-printing of the source code.
Replace every occurence of a entry like that with:

    &man.foobar.1;

Adjusting the manual page name and section number appropriately.

The definitions for these entities are stored in man-refs.ent. This
file is in doc/share/sgml because it is not just specific to the Handbook.
I expect the DocBook'd FAQ and Tutorials (coming RSN) to use them as
well.

A new PUBLIC identifier has been created for these entities, and added to
the catalog file.


git-svn-id: https://svn.freebsd.org/doc/head@4471 c2e8774f-c49f-e111-b436-862b2bbc8956