Try to solve a chicken-egg problem:
Makefile first includes Makefile.gen and then
regenerate Makefile.gen - but it don't read Makefile.gen again.
If the contents of the old and the new Makefile.gen differs
make may fail.

This happens if a category disappeared from INDEX (e.g. a typo in
CATEGORY variable). The Makefile try to build the old
category.sgml file which is no longer available.


git-svn-id: https://svn.freebsd.org/doc/www@4892 c2e8774f-c49f-e111-b436-862b2bbc8956