New make(1) infrastructure.  Split out docproj.docbook.mk into a number of smaller doc.<foo>.mk files, reflecting the functionality they contain. Long overdue, kudos to the submitter for the carrying out the work.
Also makes the files independent of the system include files that
normally live in /usr/share/mk, making it easier for non-FreeBSD systems
to download and build our docs (an important factor in making it easier
to share our work with other projects).

Finally, it (in theory) lets you build the docs with a r/o doc/ directory.

Changes to the submitted files:

doc.docbook.mk    The HTML generation depends on ${DSLHTML}, and the
                  print generation depends on ${DSLPRINT}.  Changing
                  these files will force a rebuild (which makes testing
                  changes a little easier).

                  Removed ${DOC}.doc target.  It's a hangover from when
                  I (mistakenly) thought that Jade could generate MS Word
                  .doc files.

                  Added support for using compress(1) to build .Z files
                  (been on my todo list for ages).

                  Fixed a couple of typos.

Submitted by:   Neil Blakey-Milner <nbm@mithrandr.moria.org>


git-svn-id: https://svn.freebsd.org/doc/head@5500 c2e8774f-c49f-e111-b436-862b2bbc8956