Move the discussion about CVSup out of the cutting-edge chapter and into the "Obtaining FreeBSD" appendix.
This is a straight cut-n-paste of the affected information.  However, in
its original location it was "<sect 2>", now it's "<sect1>", so all the
sectional elements have to upgraded by one.

A few references to "cvsup" as a link target were updated to "mirrors-cvsup"
as well.

Still based on work submitted by Eric Ogren, just broken down a little more.


git-svn-id: https://svn.freebsd.org/doc/head@7320 c2e8774f-c49f-e111-b436-862b2bbc8956