doc.images.mk
LOCAL_LIB_IMAGES_DIR should be a path component, not a complete path, so
remove ${.CURDIR}.

doc.docbook.mk

Set the directory for image installation correctly, and ensure that the
directory exists before we try and do anything with it.

These should fix the installation problems people are having with the
primer.  There's still an outstanding bug -- make(1) thinks that the
local library images are out-of-date with respect to the ones in
share/images for some reason.  This forces a rebuild each time.  I'm
still looking at that.


git-svn-id: https://svn.freebsd.org/doc/head@7664 c2e8774f-c49f-e111-b436-862b2bbc8956