Put down the crack pipe long enough to revert back to labelling questions with a number, and not just a big "Q".  No idea what the hell I was thinking when I did that.
Override generate-anchor in the master stylesheets, so that the names
(i.e., the <a name="...">) in the HTML output for questions is Qx.y.,
where x.y is the question number.  This is much less likely to change
than the AENxxxx format, which changes with every addition or deletion
to the FAQ, so it's much easier for people to link to specific questions
within the FAQ now.

A few whitespace/indentation changes with existing code.


git-svn-id: https://svn.freebsd.org/doc/head@8226 c2e8774f-c49f-e111-b436-862b2bbc8956