Add a new rule to the committers: respect other contributors as well.
There has been plenty of discussion about how to actually ratify
such rules, but nobody has opposed this specific rule (which is a
good show of maturity btw).

In the end the conclusion is that this is not really any different from
the rest of the tree: If it makes sense: commit it, if it creats an
uproar, back it out and paint the bikeshed.

QED.

PR:		25783
Submitted by:	Dan Langille <dan@langille.org>
Reviewed by:	developers


git-svn-id: https://svn.freebsd.org/doc/head@9084 c2e8774f-c49f-e111-b436-862b2bbc8956