Correctly display local file links in the print stylesheet.
Local-file URLs are encoded as file://localhost/path/file which is
fine for HTML but this looks ugly when printed.

Add a function to chop off the "file://localhost" so that the above
URL would be printed as "/path/file" but still link to
"file://localhost/path/file" for HTML output.


git-svn-id: https://svn.freebsd.org/doc/head@9980 c2e8774f-c49f-e111-b436-862b2bbc8956