Close pr/32278, my last submitted PR as a non-committer.
Add descriptions of what a port distfile is, and how it's used.  Use
the word distfile instead of "the program source code.

Not certain if this commit message is formatted correctly: the
X-forwarded emacs doesn't present the standard commit template.  Must
ask my mentor why not, but don't want to leave this hanging.

PR:	docs/32278


git-svn-id: https://svn.freebsd.org/doc/head@11778 c2e8774f-c49f-e111-b436-862b2bbc8956