Fix up URL references to the Handbook in the 4.6-RELEASE and 5.0-DP1 release documentation.
With one exception (the 4.6-RELEASE announce.sgml), these are all
generated files.  Usually I would be opposed to editing the generated
output, except that we're already getting bug reports resulting from
changing the way we refer to the Handbook.

Only doing this for 4.6-RELEASE and 5.0-DP1 now because these are the
latest active "releases".

Suggested by:	nik


git-svn-id: https://svn.freebsd.org/doc/www@13520 c2e8774f-c49f-e111-b436-862b2bbc8956