Re-enable the web-based PR submission mechanism.
This code relies on the netpbm port being installed ; this is
already a dependency of the textproc/docproj port.

This comes courtesy of a collaboration with Eric Anderson
<anderson@centtech.com> who submitted both the idea and the vast
majority of the code to implement it; thank you very much, Eric - this
is a great Christmas present to us and our user community.


git-svn-id: https://svn.freebsd.org/doc/www@19185 c2e8774f-c49f-e111-b436-862b2bbc8956