"Why isn't cvsup.freebsd.org a round robin DNS entry?"
If this question is asked one more time, I'm going to go bonkers with
an axe at some user group meeting.

As a Google search doesn't give much of a definitive answer from any
of the CVSup gurus, because it's drowning in the number of people
asking this question, basing this on my own knowledge.  Feel free to
expand this.


git-svn-id: https://svn.freebsd.org/doc/head@19888 c2e8774f-c49f-e111-b436-862b2bbc8956