Use a more precise ulink and prevent a needless footnote from being created.
This could also have been done with

<ulink url="&url.base;/docs.html"></ulink>.

but that would have caused "../../../docs.html" to be dislpayed in the
HTML version.  That's what we want to link to, but it looks better to
display the absolute URL inline.  This commit gives us the best of
both worlds, for print and HTML output :

          browse the list of tutorials at <ulink
-         url="&url.base;/index.html">http://www.FreeBSD.org/</ulink>.
+         url="&url.base;/docs.html">http://www.FreeBSD.org/docs.html</ulink>

Found by:	loader@freebsdmall.com


git-svn-id: https://svn.freebsd.org/doc/head@22104 c2e8774f-c49f-e111-b436-862b2bbc8956