Remove the "Does FreeBSD support any USB mice?" entry, since it describes a procedure needed only for FreeBSD 3.X.
Keeping the entry around has the risk that people do not notice that
it's not for recent FreeBSD versions, and they then get in trouble by
using this FAQ entry.

The FAQ in general still says it supports 3.X, but if anybody really
still use 3.X, I seriously doubt it's in an environment where a mouse
is required.

PR:	docs/36154


git-svn-id: https://svn.freebsd.org/doc/head@23917 c2e8774f-c49f-e111-b436-862b2bbc8956