The FAQ is guilty of egregious and annoying use of the word "simply," i.e.:
"Simply use this command" is vaguely patronizing, especially when the
command is a Perl regex with a few obscure (to new users) command-line
options attached.  "Use this command" answers the question, and
doesn't imply that you should know this already.

In most cases, when you begin a sentence with the word "simply" you
shouldn't.  Simply begin your sentence with the second word instead.


git-svn-id: https://svn.freebsd.org/doc/head@24520 c2e8774f-c49f-e111-b436-862b2bbc8956