Introduce a new entity, cgibase, which points at where our CGI scripts can be found, i.e., http://www.FreeBSD.org/cgi at the moment.
This allows for easier pre-commit testing of CGI scripts, and avoids
SNAFUs with an incorrect URL being committed, as has happened a couple
of times recently.

en/ should be fully converted with this commit; other languages need to
be updated, but are in mid merge molasses at present so I have not
touched them.


git-svn-id: https://svn.freebsd.org/doc/www@25915 c2e8774f-c49f-e111-b436-862b2bbc8956