Update of the wireless networking section:
The old section is completely replaced with this new one.
For the moment "FreeBSD as AP" is not covered (I'm working on it), but
FreeBSD clients configuration and troubleshooting are detailled (how
to find AP, basic settings with or without DHCP, authentication
methods: WEP, WPA-PSK, WPA-TLS, WPA-TTLS and PEAP, IBSS mode is also
discussed).

This huge section would not exist without the help of
loader@freebsdmall.com, murray@ and of course sam@


git-svn-id: https://svn.freebsd.org/doc/head@28395 c2e8774f-c49f-e111-b436-862b2bbc8956