Import a new translation of our Handbook, to the Mongolian language.
The work, for this FreeBSD Handbook translation, has been done by:

  * Bayanzul Lodoisamba, baynaa@yahoo.com
  * Ganbold Tsagaankhuu, ganbold@micom.mng.net
  * Natsagdorj Shagdar, natsag2000@yahoo.com

More documents are in the works, by the Mongolian FreeBSD doc
team, and Ganbold will be acting as a liaison between them and
the CVS tree for now.

Note that I haven't attached this to the toplevel doc/Makefile
yet, to make sure that it builds correctly on the FreeBSD cluster
before I make it auto-build by our doc tinderboxen.

Submitted by:   Ganbold Tsagaankhuu


git-svn-id: https://svn.freebsd.org/doc/head@29997 c2e8774f-c49f-e111-b436-862b2bbc8956