Add tab expansion procedure into ($verbatim-display$).  Although DSSSL specification supports expand-tabs: characteristics in paragraph flow objects, printable format backends in our DSSSL processors (Jade or OpenJade) do not support it.
The committed code expands a literal tab character into literal spaces with
length of %default-tab-spacing%.  In the position calculation, it assumes all
of characters have a fixed width and the number of characters in a node list
can be obtained by (string-length (data nl)).  Note that this can be false
especially when different font sizes are used.

Tested by:	wblock


git-svn-id: https://svn.freebsd.org/doc/head@38334 c2e8774f-c49f-e111-b436-862b2bbc8956