Dear Community, readers, fellow-committers,
It is my great pleasure, together with Glen, to announce that we think that
Rene is ready to fly solo. After contributing and leading the doc-nl
translations he is also perfectly capable of contributing to other
documentation area's as he demonstrated to us the last period of time.

This means we will release Rene and let him earn his own pointyhats :-)

Approved by:	doceng (gjb)


git-svn-id: https://svn.freebsd.org/doc/svnadmin@43215 c2e8774f-c49f-e111-b436-862b2bbc8956