PR:     www/159291
As remko@ assign this to me:
Implement a quick-and-dirty solution to fix a broken link on the Spanish
version of the website and link to the Spanish send-pr.html.

The Spanish version of the website has not seen an update since 2010
therefore most changes since then are either not available or are broken.

The send-pr web form is still working though.  Tested with PR www/187222.


git-svn-id: https://svn.freebsd.org/doc/head@44110 c2e8774f-c49f-e111-b436-862b2bbc8956