Update to r44080: Remove NAT section from advanced-networking chapter.
It was linked to from the firewalls chapter, but we aren't there yet.
To make the handbook build in the german tree and allow further updates
to the advanced-networking chapter, simply remove the link. The firewalls
chapter will be updated eventually, and will then get the NAT info back.

Submitted by:	    Bjoern Heidotting (changes to advanced-networking)
Obtained from:	    The FreeBSD German Documentation Project


git-svn-id: https://svn.freebsd.org/doc/head@44932 c2e8774f-c49f-e111-b436-862b2bbc8956