This is no general 'mh' port. There is no only japanese/mh.
The article references a number of old aspects include, but not limited to pkg_*, incorrect package names on the CDROM, and books or links which no longer exist.

Finally, there is not much unique to MH on FreeBSD and we do not host tutorials for all software.

Discussed with:	jmg


git-svn-id: https://svn.freebsd.org/doc/head@45155 c2e8774f-c49f-e111-b436-862b2bbc8956