Fix direct links to the clang and libxo reports
Newlines within the <title> element are retained and processed into
HTML-encoded newlines in the anchor ids, which do not function as
desired.  Thanks to Mark Eichin for the diagnosis.

While here, un-wrap a couple of <url> elements, as the hover-text
renders poorly when newlines are included there as well.

Approved by:	hrs (mentor, blanket)


git-svn-id: https://svn.freebsd.org/doc/head@46217 c2e8774f-c49f-e111-b436-862b2bbc8956