Please welcome Anne Dickison (anne@) as our newest doc committer.
Anne is the Marketing Director for the FreeBSD Foundation, and
extremely involved in the FreeBSD Project in her role.  Anne will
be helping keep the FreeBSD Project website up-to-date, primarily
in the FreeBSD advocacy areas.

I (gjb) will be Anne's mentor.

Approved by:	doceng (implicit)
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/doc/svnadmin@46871 c2e8774f-c49f-e111-b436-862b2bbc8956