Update to r49828:
Shuffle some text in the ports section of the Handbook.

The presentation assumes more knowledge on the part of a new user to
FreeBSD than it should.

In good writing, concepts should always be introduced before being
discussed. The current presentation mixes them. The new order introduces
applications, then ports, then packages, and then elaborates on the
difference between packages and ports.

90% of the content is merely moved. Some editing for clarity is included.


git-svn-id: https://svn.freebsd.org/doc/head@49921 c2e8774f-c49f-e111-b436-862b2bbc8956