Update to r50777:
Subversion has been told to expand $FreeBSD$ version strings in source
files. This is good, evidently.

But we don't want it to expand version strings in examples, because then
the example will have an expanded version string with the version of the
current doc file, which is bad.  The version number, not the doc file.
Or, well, maybe both, but the point here is the version number won't look
right.

So use the good and righteous &dollar;FreeBSD&dollar; construct instead.
This is perfectly cromulent and there is nothing remotely ridiculous or
surreal about this at all.  Lalala, I can't hear you.


git-svn-id: https://svn.freebsd.org/doc/head@50786 c2e8774f-c49f-e111-b436-862b2bbc8956