Update to r50779:
Add its:translate="no" properties to elements containing $FreeBSD$
version strings.  This will prevent them from showing up in PO
translations and causing difficulties.

It has the unfortunate side effect that the entire element contents do
not appear in PO files.  Most of these are elements are <programlisting>,
which are probably not translated often yet. At least one is a <screen>,
which might be more of a problem.

At this point, this is a workaround to make it possible for PO translators
to continue work without the ever- changing version strings creating
conflicts.


git-svn-id: https://svn.freebsd.org/doc/head@50864 c2e8774f-c49f-e111-b436-862b2bbc8956