Do not set NO_OBJ=YES in a Makefile because /usr/share/mk/bsd.obj.mk does a test if NO_OBJ is defined, but not about the value.
Which means you cannot unset NO_OBJ on the command line or with env variable:

$ make NO_OBJ=""

or

$ export NO_OBJ=""; make

I still think that 'make obj' doesn't really work for the doc
project. But for now, you have to disable it yourself on the command
line or in your ~/.profile.

This is a revert of r51065


git-svn-id: https://svn.freebsd.org/doc/head@51278 c2e8774f-c49f-e111-b436-862b2bbc8956