Update to r51447:
Fix command to extract multiple sets for jail creation

In the example for extracting multiple distribution sets the path and
case was incorrect.

There's still more adjustment to be made here: not all sets are listed,
and one of the unlisted ones (lib32) does not exist on all archs, but
at least the example command should work.


git-svn-id: https://svn.freebsd.org/doc/head@51624 c2e8774f-c49f-e111-b436-862b2bbc8956