Revert r51644: restore Open Search
I still can't seem to get this to work on either Firefox or Chrome, but
it is likely that I'm testing wrong rather than finding a non-working
feature.

This leaves the actually broken files removed from earlier commits, but
restores the ones that should still be working.

Reported by:	jhb


git-svn-id: https://svn.freebsd.org/doc/head@51713 c2e8774f-c49f-e111-b436-862b2bbc8956