Fix broken 'uses-values' links in the Porter's Handbook
The USES chapter sets an xml:id on its title tag but this doesn't
seem to generate an anchor.  As a consequence there currently are
several 'uses-values' xrefs that point to nowhere.

This is only really noticeable in the single page HTML version
(book.html) of the handbook.

PR:		227123
Reported by:	0mp
Approved by:	mat
Differential Revision:	https://reviews.freebsd.org/D15642


git-svn-id: https://svn.freebsd.org/doc/head@51756 c2e8774f-c49f-e111-b436-862b2bbc8956