Add more information about Bluetooth usage.
The patch shows how to create a remote device connection and
how to manage devices in /etc/bluetooth/hosts.  The multimedia
chapter was extended with a description on how to set up and
connect a Bluetooth audio device as a sound sink.
It also adds man page references to man-refs.ent for hcsecd(8)
and virtual_oss(8), respectively.

I cleaned up the patch a bit for consistency and proper markup.

PR:			    228879
Submitted by:		    nekoxmachina@gmail.com
Take a look at this PR by:  lme


git-svn-id: https://svn.freebsd.org/doc/head@52744 c2e8774f-c49f-e111-b436-862b2bbc8956