Update the Xen chapter to support newer versions of Xen.
This patch clarifies which versions of FreeBSD support Xen 4.7 (PVHv1)
and Xen 4.11 or greater (PVHv2). Additionally, improve a few of the
instructions. Use onifconsole for Xen, add a line to loader.conf to
load the tap device. Remove restrictions to use MMCFG PCI accesses for
Xen >= 4.11. Make /var/log/xen stand out a bit more by making it a tip.
Finally, the example guest was creating a 10.3 version of FreeBSD,
which is no longer supported. Update that to 12.0 and wrap it in
replaceable tags so that people know they can put something else there.

Big thanks to royger@ for the updates and doing the documentation work
before I had a chance to do it.

Submitted by:		royger@
Reported by:		freebsd-xen mailinglist
Approved by:		bcr@
Differential Revision:	https://reviews.freebsd.org/D19106


git-svn-id: https://svn.freebsd.org/doc/head@52813 c2e8774f-c49f-e111-b436-862b2bbc8956