Add an entry to the administration page for the ports security team.
The ports security teams exists for a while now, so it should have a
proper listing on the administration page with all current members.

Thanks to Kurt Jaeger for the writeup and submitting the patch.
I only rewrapped one line to fit within the line limits.

PR:		    235939
Submitted by:	    pi@


git-svn-id: https://svn.freebsd.org/doc/head@52850 c2e8774f-c49f-e111-b436-862b2bbc8956