Ensure the generated ports category XML files are always recreated.
If a build failure is triggered, such as a port Makefile
contains a non-ASCII apostrophe, the generated files are
not replaced if they exist (even if an error occurs).
In such a case, manual intervention would be needed to
remove the file(s) that would cause the build failure.

Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/doc/head@52882 c2e8774f-c49f-e111-b436-862b2bbc8956