Update to r52814:
Add a troubleshooting section to the Xen chapter.

In the discussion leading up to r52813, rgrimes@ suggested to add a
troubleshooting section to the Xen chapter to help with debugging issues.

In particular, users should know where Xen 4.11 and higher logs important
information relevant for debugging. There are also a couple of options
that increase the verbosity of the dom0 and enable iommu debugging. These
were divided up into two subsections: one for the host and the other for
the guest.


git-svn-id: https://svn.freebsd.org/doc/head@53004 c2e8774f-c49f-e111-b436-862b2bbc8956