Rewrite the platform Tier definitions.
Rewrite the prose description of Tiers to be structured as bullet
lists of guarantees to users from the Project, guarantees to
developers from the Project, and obligations on developers.

This includes definitions of userland and kernel ABIs as well as
documenting our current practice of ABI stability.

The committments for ports are still vague and will require further
refinement.

Move the Tier status of architectures out of the committers guide and
into a table on the platforms page the website listing the Tier for
each architecture across currently supported stable branches as well
as the projected Tiers for the next stable branch (in this case,
13.x).  The table also lists individual TARGET_ARCH values to permit
more granularity in Tier definitions (e.g. hard-float vs soft-float).

Update the Unsupported Platforms table to only list removed
architectures and include the last supported release of these
architectures.  This required adding anchors for relevant releases on
the releases page.

Reviewed by:	bcr
Discussed with:	developers@
Differential Revision:	https://reviews.freebsd.org/D22439


git-svn-id: https://svn.freebsd.org/doc/head@53619 c2e8774f-c49f-e111-b436-862b2bbc8956