[PowerPC] Update /platforms to reflect the powerpc* switch to clang.
As of r356111, the default compiler on all powerpc* targets in HEAD is
no longer GCC4.

As such, remove powerpc* from the automatic-transition-to-tier-4 list.

(This includes powerpcspe, which is currently broken as-is in HEAD because
of some target triple issues that will be fixed soon, but builds and works
fine in clang with some minor out-of-tree patches.)

Approved by:	docs
Differential Revision:	https://reviews.freebsd.org/D22934


git-svn-id: https://svn.freebsd.org/doc/head@53710 c2e8774f-c49f-e111-b436-862b2bbc8956