Add optional fop install step to generate PDF output.
The textproc/fop is an optional package that can be installed to generate
PDF output of our docs. It is not included in textproc/docproj by default
and if not strictly required, but for those who want PDFs we describe how.

While here, bump the copyright dates of the FDP primer.

Thanks to Pau Amma for this change.

PR:		247019
Submitted by:	pauamma_gundo.com
Reviewed by:	debdrup, allanjude
Approved by:	allanjude, bcr
Differential Revision:	https://reviews.freebsd.org/D25170


git-svn-id: https://svn.freebsd.org/doc/head@54256 c2e8774f-c49f-e111-b436-862b2bbc8956