Don't recommend the -r flag to mountd.
It enables a non-standards compliant feature to support ancient
diskless terminals.  We shouldn't recommend it without explanation.
(-r is currently the default in /etc/defaults/rc.conf along with -S,
this should likely be revisited.)

Instead, suggest setting mountd_enable.  While not required to start
the NFS server, it is required for the "service mountd restart" command
below.

Reviewed by:	bcr
Differential Revision:	https://reviews.freebsd.org/D25760


git-svn-id: https://svn.freebsd.org/doc/head@54368 c2e8774f-c49f-e111-b436-862b2bbc8956